+++ CMakeLists.txt
@@ -193,7 +193,6 @@ add_library(algo STATIC
     src/algo/Graph.cpp
     src/algo/GraphBuilder.cpp
     src/algo/Scoring.cpp
-    src/algo/BetweennessCentrality.cpp
     src/algo/ReadingSequencer.cpp
     src/algo/AlgoRunner.cpp
     src/algo/AlgoDbWriter.cpp
